requirements.txt
numpy~=2.0.2
numba~=0.60.0
polars~=1.2.0
pillow~=10.4.0
pandas~=2.2.2
scipy~=1.13.1
seaborn~=0.13.2
matplotlib
statsmodels